test: add get keys of object
