+++ types/src/GetKeysOfObjectAsUnion.ts
@@ -19,37 +19,3 @@ export type FlattenKeys<T extends object> = {
 export type GetKeysOfObjectAsUnion<T extends object> = FlattenKeys<PrefixKeysDeep<T>>;
 
 export default GetKeysOfObjectAsUnion;
-
-// test interface
-interface Test {
-  id: number;
-  name?: string;
-  address: {
-    id: number;
-    index: number;
-    street: string;
-    city: string;
-    isLiving: boolean;
-  };
-  work: {
-    name: string;
-    place: string;
-    department: {
-      depId: number;
-      depName: string;
-    }
-  }
-}
-
-// test array
-const keys: Array<GetKeysOfObjectAsUnion<Test>> = [
-  'name',
-  'address',
-  'nonExists', // error
-  'address.index',
-  'address.isLiving',
-  'address.nonExists', // error
-  'work.name',
-  'work.department.depId',
-  'work.department.id', // error
-];
